+++ common/lib/yandex-metrika-script-url.d.ts
@@ -1,3 +1,0 @@
-declare function buildYandexMetrikaScriptUrl(options?: { assetVersion?: number; isDev?: boolean }): string;
-
-export { buildYandexMetrikaScriptUrl };