+++ common/templates/client-entry-script.d.ts
@@ -1,0 +1,5 @@
+/// <reference path="../../types/index.d.ts" />
+
+declare function renderClientEntryScript(options?: ClientEntryScriptOptions): string;
+
+export { renderClientEntryScript };