+++ common/templates/render-inline-page-assets.js
@@ -1,6 +1,5 @@
 import { getSiteConfig } from '#core/common/lib/site-config.js';
 import { renderPageAssets } from '#core/common/templates/page-assets.js';
-import { renderYandexMetrikaInlineHead } from '#core/common/templates/yandex-metrika-inline.js';
 import { getCss, getJs } from '#core/server/lib/inline-bundle.js';
 
 /**
@@ -11,7 +10,6 @@ import { getCss, getJs } from '#core/server/lib/inline-bundle.js';
 function createRenderInlinePageAssets({
 	cssEntry = 'main.css',
 	cwd = process.cwd(),
-	includeInlineMetrika = true,
 	jsEntry = 'entries/main.js',
 } = {}) {
 	let cssCache = '';
@@ -32,14 +30,9 @@ function createRenderInlinePageAssets({
 			jsCache = await getJs(jsEntry, { cwd });
 		}
 
-		const metrikaTemplate = includeInlineMetrika
-			? renderYandexMetrikaInlineHead({ pathname: pageAssetsOptions.pathname })
-			: '';
-
 		return /* html */ `
 			<style>${cssCache}</style>
 			<script type="module">${jsCache}</script>
-			${metrikaTemplate}
 		`;
 	};
 }