+++ common/templates/yandex-metrika-script.d.ts
@@ -1,5 +1,0 @@
-/// <reference path="../../types/index.d.ts" />
-
-declare function renderYandexMetrikaScript(options?: YandexMetrikaScriptOptions): string;
-
-export { renderYandexMetrikaScript };